requirements.txt
djangorestframework
django-crispy-forms
django-heroku
gunicorn
psycopg2
requests
django
bs4
jwt